requirements.txt
pandas==2.3.3
requests==2.32.5
python-dotenv==1.1.0
